service/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
numpy>=1.24.0
torch>=2.1.0
torchaudio>=2.1.0
librosa>=0.10.1
soundfile>=0.12.1
requests>=2.31.0
python-dotenv>=1.0.0
chatterbox-tts>=0.1.0